refactored actions into folder
