+++ src/actions/modal.js
@@ -1,0 +1,7 @@
+import { TOGGLE_REQUEST_MODAL } from '../constants'
+
+export function toggleRequestModal() {
+  return {
+    type: TOGGLE_REQUEST_MODAL
+  }
+}
