+++ src/constants.js
@@ -1,0 +1,2 @@
+export const TOGGLE_REQUEST_MODAL = 'TOGGLE_REQUEST_MODAL'
+export const INIT_NPS_CAMPSITES = 'INIT_NPS_CAMPSITES'
