+++ src/data/defaultModalState.js
@@ -1,0 +1,3 @@
+export default {
+  isModalOpen: false,
+}
